‎

+++ String/suffixAutomaton1.cpp
@@ -8,7 +8,7 @@ namespace sam {
     };
     int cur, sz;
     vector<node> sa(MAX * 2);
- 
+
     void add(char c) {
         int at = cur;
         sa[cur].fpos = sa[sz].len = sa[cur].len + 1;
@@ -23,7 +23,7 @@ namespace sam {
         while (at != -1 && sa[at].next[c] == q) sa[at].next[c] = qq, at = sa[at].link;
         sa[q].link = sa[cur].link = qq;
     }
- 
+
     void build(string& s) {
         #warning "clear????";
         sa.assign(MAX * 2, node());
@@ -33,7 +33,7 @@ namespace sam {
         while (at) sa[at].acc = 1, at = sa[at].link;
     }
     int64_t distinct_substrings() {
-        ll ans = 0;
+        int ans = 0;
         for (int i = 1; i < sz; i++) ans += sa[i].len - sa[sa[i].link].len;
         return ans;
     }
@@ -61,14 +61,14 @@ namespace sam {
         for (int i = MAX - 1; i > 0; --i) match[i] = max(match[i], match[sa[i].link]);
         for (int i = 0; i < MAX; ++i) LCS[i] = min(LCS[i], match[i]);
     }
- 
+
     int lcs_n(vector<string>& t) {
         const int INF = 1e7;
         LCS.assign(MAX, INF);
         forn(i, sz(t)) lcsMatch(t[i]);
         return *max_element(all(LCS));
     }
- 
+
     int isSubstr(string& s) {
         int at = 0;
         for (auto& i : s) {
@@ -77,21 +77,21 @@ namespace sam {
         }
         return at;
     }
- 
+
     int count_occ(int u) {
         if (sa[u].cnt != 0) return sa[u].cnt;
         sa[u].cnt = sa[u].acc;
         for (auto& v : sa[u].next) sa[u].cnt += count_occ(v.s);
         return sa[u].cnt;
     }
- 
+
     int pos_occ(string& s) {
         int x = sam::isSubstr(s);
         return x ? (abs(sam::sa[x].fpos - sz(s)) + 1) : -1;
     }
- 
-    ll dp[2 * MAX];
-    ll paths(int i) {
+
+    int dp[2 * MAX];
+    int paths(int i) {
         auto& x = dp[i];
         if (x) return x;
         x = 1;
@@ -100,7 +100,7 @@ namespace sam {
         }
         return x;
     }
- 
+
     void kth_substring(int k, int at = 0) { // k=1 : menor substring lexicog.
         for (int i = 0; i < 26; i++) if (k && sa[at].next.count(i + 'a')) {
             if (paths(sa[at].next[i + 'a']) >= k) {
